Finishing up writing tests.
This commit finishes up writing the desired tests for this module. Some changes were also made to render sender based on the results of the tests. The tests no longer
use the npm package stream-equal. The tests no use sha hashes to see if the rendering happened correctly or not.