Removing size.
This commit is removing size from the opts. The ffmpeg screenshot uses dimensions for the size. The renderSend function now sets dimensions to width by height if they
exist. This way width and height can given or dimensions can be given and it will work properly. An example url would be:

https://www.paviliongift.com/item-image/fileName-100x100-timestamp:1.mp4.jpg

This will create an image of size 100 by 100 from the first second of the video. The url could also be the same, but instead -width:100 -height:100 would work as well.
This commit also creates a new test for testing the usage of width and height.