Memory Leak and file ext.
This commit works on fixing a memory leak error and the service responding with the file type isn't supported. The memory leak was due to the fact the
read streams where not being closed before the cb was being invoked. There good also be some other issues with memory leaking going on. This commit also
checks the lowercase version of the file extension just incase the file extension is in caps. This prevents the unwated behavior of the application saying
the file type is not supported for '.JPG' when it should be. A test still needs to be written for this case.